Replace dot2any.py with a shell wrapper script
Using Python to wrap shell commands seems to be overkill.
The code and commands are simplified
to a convenience utility shell script.
Recursive conversion of dot files within sub-directories
can be done with xargs.

The functionality doesn't look to be a part of cppdep,
but it is an interesting idea to incorporate this option
into ``cppdep.py`` script directly.
That is, instead of outputting 'dot' files with `write_dot`,
cppdep could generate image/pdf files directly (e.g., write_pdf).
However, this will add Graphviz as a mandatory requirement.