Remove code dealing with external hbase conflicts
The implementation with base-name based lookup is a hacky optimization
to avoid following header files from context dependent search.
The side-effect of this optimization
is the internal and external file name conflicts
even though these files belong to completely different packages.

The optimization is presented as a bogus feature in cppdep.
Other tools can do duplicate detection much better and more focused
than cppdep should do,
for this concern is not relevant to physical dependency analysis.