Refactor and fix CCD calculation implementation
The implementation of the CCD calculation is moved into Graph class.
Double counting of common components is fixed.
This fixes #11.

The code and future for layering calculation is removed
as for the bug #9.
The tests for the old API are skipped for now
to be enabled again after level calculation is implemented.