Merge ldep and cdep into analyze()
The components are identified early without a need for late ldep.
The dependency graph already encodes the information for ldep,
so chasing header includes is unnecessary.

The current refactoring is radical.
It breaks ldep functionality,
the performance suffers by 5 - 10x.

This change is necessary to make header lookup context aware.

Issue #6.