Use objects in graph nodes instead of string ids
The objects are used directly in dependency graph nodes
instead of mapping back-n-forth with ids.
The objects provide string conversion operator
for printing and sorting.

The reason to use objects directly
is to remove the extra indirection to the object information.
Now the graph does not need to know
where the objects are coming from
to search for the objects with their ids.
The graph has a direct access
to all the information the object provides.
This is both convenient and efficient.

Moreover, hashing pointers should be more efficient than strings.
However, it is not showing up in profiling.