Implement package src/include/alias paths
Instead of having a single path container,
a project is defined by different paths
with different functionalities.

This approach is similar to the original dep_utils.
src path points to a directory with source files to process.
include path provides a directory to be added into header search.
alias path associates a directory with the package.

The alias path is not fully implemented to associate
external sub-packages within a group.