Remove implicit single alias Package construction
Explicit is better than implicit!
This should be handled outside of package constructor.
The implicit single source packages
are not supported by YAML config (#24),
so the code is unused.

Some ugly code due to this initial mal-design is removed.
This feature could be reintroduced in future in a better form.