Handle header files w/o extensions
The regex for source file search is updated
to work with files without extensions in the source directories.
These extension-less files are assumed to be header files;
however, this may lead to many false positives,
requiring either 'ignore' or 'optional' features.

In order to support the extension-less headers,
the regex ignores files with '.' and ' ' (+ other special chars)
in the filename.
It is unconventional to use special chars in source file names.

Fixes #32