Accept glob pattern for src paths
glob pattern somewhat complicates
finding a common path for a set of paths.
In addition, verification whether the path is directory is given up
because the source path can also be a file.
Moreover, since files cannot be a package alias,
alias paths are now only associated with include semantics
(only used for external packages).

The error case of a failure to associate a header to a component
is presented as Analysis configuration error for the user
due to the missing information or misinformation in the configuration.

Closes #36