Implement ignore/exclude paths from source paths
This is similar to the original adep's ignore path feature,
but this implementation works with glob patterns.

Note that the performance cost is considerable
because the ignored path patterns are not subtracted
from the source path patterns;
instead, every path upon os directory walk
is checked against the exclusion patterns,
so the complexity ends up being O(N*M).

Issue #36.
Closes #23.