Pair header and impl files in different dirs
This is the initial implementation of an algorithm
to pair header and implementation files in different directories.
The expected/average pairing complexity
is changed from O(N) to O(N*logN).
The worst case is O(N^2).

Moreover, the algorithm can result
in false positive pairing of unrelated header and implementation files
within a single project.

The approach is based on heuristics
that header and implementation file locations
have similar directory structures,
so this is a general case of when these files are in the same directory.
The basic idea is to sort and group candidate files
by common directory structure.
Note that these candidate files have the same filename w/o extensions.
The heuristics pairs the files
with the most number of common ancestor directories
starting from the files (not the root directory!).

The implementation went through minimal (manual) testing,
and it should get tested on real projects.

Closes #19.